pyproject.toml
[project]
name = "llm-bench"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.28.1",
    "matplotlib>=3.9.4",
    "mlx-lm>=0.29.1",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "seaborn>=0.13.2",
]
